api/requirements.txt
pillow
fastapi
uvicorn
onnxruntime-gpu
pydantic
